Rename async_functions to asyncio_extensions
